requirements.txt
pyserial == 3.5
# PyGObject == 3.38.0
dbus-python == 1.2.18
pyusb == 1.2.0
gatt == 0.2.7
supervisor == 4.2.1
luma.oled == 3.8.1
paho-mqtt==2.1.0
python-dotenv==1.0.1